use uint64 when creating bitmasks

+++ lib/util/bit_set.py
@@ -213,7 +213,7 @@ class BitSet:
 
         word_index = self._word_index(bit_index)
         self._expand_to(word_index)
-        self._words[word_index] ^= circular_left_shift(1, bit_index, self.bits_per_word)
+        self._words[word_index] ^= circular_left_shift(uint64(1), bit_index, self.bits_per_word)
         
         self._recalculate_words_in_use()
         self._ensure_invariants()
@@ -233,7 +233,7 @@ class BitSet:
         
         word_index = self._word_index(bit_index)
         self._expand_to(word_index)
-        self._words[word_index] |= circular_left_shift(1, bit_index, self.bits_per_word)
+        self._words[word_index] |= circular_left_shift(uint64(1), bit_index, self.bits_per_word)
     
     def set_region(self, bit_slice: slice, value: SupportsBool = True) -> None:
         raise NotImplemented
@@ -249,7 +249,7 @@ class BitSet:
         if word_index >= self._words_in_use:
             return
 
-        self._words[word_index] &= ~circular_left_shift(1, bit_index, self.bits_per_word)
+        self._words[word_index] &= ~circular_left_shift(uint64(1), bit_index, self.bits_per_word)
         
         self._recalculate_words_in_use()
         self._ensure_invariants()
@@ -299,7 +299,7 @@ class BitSet:
         word_index = self._word_index(bit_index)
         if word_index >= self._words_in_use:
             return False
-        masked_word = self._words[word_index] & circular_left_shift(1, bit_index, self.bits_per_word)
+        masked_word = self._words[word_index] & circular_left_shift(uint64(1), bit_index, self.bits_per_word)
         return bool(masked_word != 0)
 
     def get_region(self, bit_slice: slice) -> Self:
